Tests/LibWeb/Text/expected/CSSOMView/click-focused-element-in-scroller-should-not-scroll.txt

scrollTop before click: 342
scrollTop after click: 342
